requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
build==1.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
datasets==2.17.1
dill==0.3.8
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface_hub==0.20.3
hypothesis==6.98.10
idna==3.6
Jinja2==3.1.3
kenlm==0.2.0
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
packaging==23.2
pandas==2.2.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pyctcdecode==0.5.0
pygtrie==2.5.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sentencepiece==0.2.0
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
sympy==1.12
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.0
torchaudio==2.2.0
tqdm==4.66.2
transformers==4.38.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
